requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
spacy==3.5.0
tensorflow==2.9.1
torch==1.12.1
scikit-learn==1.1.1
requests==2.28.1
pyjwt==2.4.0
cryptography==38.0.0
